+++ src/components/Home/Home.js
@@ -3,8 +3,9 @@ import { connect } from 'react-redux'
 
 /**/
 import styles from './Home.scss'
-import LeftBar from '../LeftBar'
+import LeftBar from '../LeftBar/LeftBar'
 import MainPanel from '../MainPanel'
+import Masthead from '../Masthead/Masthead'
 
 class Home extends React.Component {
 
@@ -13,11 +14,13 @@ class Home extends React.Component {
   }
   
   render () {
-
     return (
       <div className={styles.wrapper}>
-        <LeftBar className={styles.leftBar}/>
-        <MainPanel className={styles.mainPanel}/>
+        <Masthead>1</Masthead>
+        <div className={styles.body}>
+          <LeftBar className={styles.leftBar}/>
+          <MainPanel className={styles.mainPanel}/>
+        </div>
       </div>
     )
   }
