+++ src/components/LeftBar/LeftBar.scss
@@ -4,15 +4,8 @@
   height: 100%;
 
   .offset {
-    border: 1px solid yellow;
-    width: 220px;
-    height: 100%;
-  }
-
-  .content {
-    position: fixed;
-    top: 0;
-
+    height: 50px;
+    background-color: gray;
   }
 
 }
