+++ src/components/Logo/Logo.scss
@@ -1,8 +1,9 @@
 
-
 .wrapper {
-  border: 1px solid green;
+
   width: 140px;
-  height: 75px;
+  height: 50px;
+
+  background-image: url('http://www.swmaestro.kr/web/web/_Img/Common/logo.gif');
 
 }
