+++ src/components/Logo/package.json
@@ -2,5 +2,5 @@
   "name": "Logo",
   "version": "0.0.0",
   "private": true,
-  "main": "./Logo.jsx"
+  "main": "./Logo.js"
 }
